ACCUMULO-4334 JMX Ingest returns double

+++ server/tserver/src/main/java/org/apache/accumulo/tserver/metrics/TabletServerMBean.java
@@ -38,7 +38,7 @@ public interface TabletServerMBean {
 
   long getQueries();
 
-  long getIngest();
+  double getIngest();
 
   long getTotalMinorCompactions();
